utilities/make_mask_square/pyproject.toml
[tool.poetry]
name = "make_mask_square"
version = "0.1.0"
description = ""
authors = ["Kshitiz <dzgrotechnologies@gmail.com>"]
readme = "README.md"
packages = [{ include = "make_mask_square", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12"
pillow = "*"
pytest = "*"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

